+++ packages/terra-core-docs/src/terra-dev-site/test/form-radio/radio/HiddenRadio.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Radio from 'terra-form-radio';
+
+export default () => <Radio id="hidden" labelText="can you see me?" isLabelHidden />;