+++ packages/terra-core-docs/src/terra-dev-site/test/form-radio/radio/LongTextRadio.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Radio from 'terra-form-radio';
+
+export default () => (
+  <Radio
+    id="long-text"
+    labelText={'Lorem ipsgit aum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua '
+              + 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua'}
+  />
+);