+++ packages/terra-form-radio/src/terra-dev-site/test/form-radio/radio/DefaultRadio.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Radio from '../../../../Radio';
-
-export default () => <Radio id="default" labelText="Default Radio" />;