+++ packages/terra-form-radio/src/terra-dev-site/test/form-radio/radio/HiddenRadio.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Radio from '../../../../Radio';
-
-export default () => <Radio id="hidden" labelText="can you see me?" isLabelHidden />;